+++ src/Controller/InteractController.php
@@ -1,0 +1,104 @@
+<?php
+
+/* This file is part of NextDom Software.
+ *
+ * NextDom is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * NextDom Software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with NextDom Software. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @Support <https://www.nextdom.org>
+ * @Email   <admin@nextdom.org>
+ * @Authors/Contributors: Sylvaner, Byackee, cyrilphoenix71, ColonelMoutarde, edgd1er, slobberbone, Astral0, DanoneKiD
+ */
+
+namespace NextDom\Controller;
+
+use NextDom\Helpers\Status;
+use NextDom\Helpers\PagesController;
+use NextDom\Helpers\Render;
+use NextDom\Managers\CmdManager;
+use NextDom\Managers\JeeObjectManager;
+use NextDom\Managers\EqLogicManager;
+
+class InteractController extends PagesController
+{
+
+    
+    public function __construct()
+    {
+        Status::initConnectState();
+        Status::isConnectedAdminOrFail();
+    }
+
+    /**
+     * Render interact page
+     *
+     * @param Render $render Render engine
+     * @param array $pageContent Page data
+     *
+     * @return string Content of interact page
+     *
+     * @throws \NextDom\Exceptions\CoreException
+     * @throws \Twig_Error_Loader
+     * @throws \Twig_Error_Runtime
+     * @throws \Twig_Error_Syntax
+     */
+    public static function interact(Render $render, array &$pageContent): string
+    {
+
+        $interacts = array();
+        $pageContent['interactTotal'] = \interactDef::all();
+        $interacts[-1] = \interactDef::all(null);
+        $interactListGroup = \interactDef::listGroup();
+        if (is_array($interactListGroup)) {
+            foreach ($interactListGroup as $group) {
+                $interacts[$group['group']] = \interactDef::all($group['group']);
+            }
+        }
+        $pageContent['JS_END_POOL'][] = '/public/js/desktop/tools/interact.js';
+        $pageContent['interactsList'] = $interacts;
+        $pageContent['interactsListGroup'] = $interactListGroup;
+        $pageContent['interactDisabledOpacity'] = \nextdom::getConfiguration('eqLogic:style:noactive');
+        $pageContent['interactCmdType'] = \nextdom::getConfiguration('cmd:type');
+        $pageContent['interactAllUnite'] = CmdManager::allUnite();
+        $pageContent['interactJeeObjects'] = JeeObjectManager::all();
+        $pageContent['interactEqLogicTypes'] = EqLogicManager::allType();
+        $pageContent['interactEqLogics'] = EqLogicManager::all();
+        $pageContent['interactEqLogicCategories'] = \nextdom::getConfiguration('eqLogic:category');
+        $pageContent['JS_END_POOL'][] = '/public/js/adminlte/utils.js';
+
+        return $render->get('/desktop/tools/interact.html.twig', $pageContent);
+    }
+
+    /**
+     * Render interactAdmin page
+     *
+     * @param Render $render Render engine
+     * @param array $pageContent Page data
+     *
+     * @return string Content of interact_admin page
+     *
+     * @throws \NextDom\Exceptions\CoreException
+     * @throws \Twig_Error_Loader
+     * @throws \Twig_Error_Runtime
+     * @throws \Twig_Error_Syntax
+     */
+    public static function interactAdmin(Render $render, array &$pageContent): string
+    {
+
+        $pageContent['JS_END_POOL'][] = '/public/js/desktop/params/interact_admin.js';
+        $pageContent['JS_END_POOL'][] = '/public/js/adminlte/utils.js';
+
+        return $render->get('/desktop/params/interact_admin.html.twig', $pageContent);
+    }
+
+}
