+++ src/Controller/LinksController.php
@@ -25,9 +25,8 @@ namespace NextDom\Controller;
 use NextDom\Helpers\PagesController;
 use NextDom\Helpers\Render;
 use NextDom\Helpers\Status;
-use NextDom\Helpers\Utils;
 
-class RealtimeController extends PagesController
+class LinksController extends PagesController
 {
     
     public function __construct()
@@ -36,30 +35,26 @@ class RealtimeController extends PagesController
         Status::isConnectedAdminOrFail();
     }
 
+    
     /**
-     * Render realtime page
+     * Render links page
      *
      * @param Render $render Render engine
      * @param array $pageContent Page data
      *
-     * @return string Content of log_admin page
+     * @return string Content of links page
      *
      * @throws \NextDom\Exceptions\CoreException
      * @throws \Twig_Error_Loader
      * @throws \Twig_Error_Runtime
      * @throws \Twig_Error_Syntax
      */
-    public static function realtime(Render $render, array &$pageContent): string
+    public static function links(Render $render, array &$pageContent): string
     {
 
-
-        $pageContent['JS_VARS']['realtime_name'] = Utils::init('log', 'event');
-        $pageContent['JS_VARS']['log_default_search'] = Utils::init('search', '');
-
-        $pageContent['JS_END_POOL'][] = '/public/js/desktop/diagnostic/realtime.js';
+        $pageContent['JS_END_POOL'][] = '/public/js/desktop/params/links.js';
         $pageContent['JS_END_POOL'][] = '/public/js/adminlte/utils.js';
 
-        return $render->get('/desktop/diagnostic/realtime.html.twig', $pageContent);
+        return $render->get('/desktop/params/links.html.twig', $pageContent);
     }
-
 }
