+++ src/Controller/SecurityController.php
@@ -1,0 +1,88 @@
+<?php
+
+/* This file is part of NextDom Software.
+ *
+ * NextDom is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * NextDom Software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with NextDom Software. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @Support <https://www.nextdom.org>
+ * @Email   <admin@nextdom.org>
+ * @Authors/Contributors: Sylvaner, Byackee, cyrilphoenix71, ColonelMoutarde, edgd1er, slobberbone, Astral0, DanoneKiD
+ */
+
+namespace NextDom\Controller;
+
+use NextDom\Helpers\Status;
+use NextDom\Helpers\PagesController;
+use NextDom\Helpers\Render;
+use NextDom\Managers\CacheManager;
+ 
+class SecurityController extends PagesController
+{
+
+    public function __construct()
+    {
+        Status::initConnectState();
+        Status::isConnectedAdminOrFail();
+    }
+
+    /**
+     * Render security page
+     *
+     * @param Render $render Render engine
+     * @param array $pageContent Page data
+     *
+     * @return string Content of security page
+     *
+     * @throws \NextDom\Exceptions\CoreException
+     * @throws \Twig_Error_Loader
+     * @throws \Twig_Error_Runtime
+     * @throws \Twig_Error_Syntax
+     */
+    public static function security(Render $render, array &$pageContent): string
+    {
+        $keys = array('security::bantime', 'ldap::enable');
+        $configs = \config::byKeys($keys);
+
+        $pageContent['JS_VARS']['ldapEnable'] = $configs['ldap::enable'];
+
+        $pageContent['adminUseLdap'] = function_exists('ldap_connect');
+        if ($pageContent['adminUseLdap']) {
+            $pageContent['adminLdapEnabled'] = \config::byKey('ldap:enable');
+        }
+        $pageContent['adminBannedIp'] = [];
+        $cache = CacheManager::byKey('security::banip');
+        $values = json_decode($cache->getValue('[]'), true);
+
+        if (is_array($values) && count($values) > 0) {
+            foreach ($values as $value) {
+                $bannedData = [];
+                $bannedData['ip'] = $value['ip'];
+                $bannedData['startDate'] = date('Y-m-d H:i:s', $value['datetime']);
+                if ($configs['security::bantime'] < 0) {
+                    $bannedData['endDate'] = __('Jamais');
+                } else {
+                    $bannedData['endDate'] = date('Y-m-d H:i:s', $value['datetime'] + $pageContent['adminConfigs']['security::bantime']);
+                }
+                $pageContent['adminBannedIp'][] = $bannedData;
+            }
+        }
+
+        $pageContent['JS_END_POOL'][] = '/public/js/desktop/admin/security.js';
+        $pageContent['JS_END_POOL'][] = '/public/js/adminlte/utils.js';
+
+        return $render->get('/desktop/admin/security.html.twig', $pageContent);
+    }
+ 
+
+}
